requirements.txt
numpy
pandas
tensorflow-addons
opencv-python
tqdm
tensorflow-gpu
matplotlib
pytest
jupyter
openpyxl